refactor: Move a test helper function from `core` to `ktor-utils`
This makes the function available for tests in components.
